+++ app/celer-g4/RunInputIO.json.cc
@@ -13,28 +13,10 @@
 #include "corecel/io/StringEnumMapper.hh"
 #include "corecel/sys/Environment.hh"
 #include "celeritas/Types.hh"
+#include "celeritas/TypesIO.hh"
 #include "celeritas/ext/GeantPhysicsOptionsIO.json.hh"
 #include "celeritas/field/FieldDriverOptionsIO.json.hh"
 #include "celeritas/phys/PrimaryGeneratorOptionsIO.json.hh"
-
-namespace celeritas
-{
-//---------------------------------------------------------------------------//
-void from_json(nlohmann::json const& j, TrackOrder& value)
-{
-    static auto const from_string
-        = StringEnumMapper<TrackOrder>::from_cstring_func(to_cstring,
-                                                          "track order");
-    value = from_string(j.get<std::string>());
-}
-
-void to_json(nlohmann::json& j, TrackOrder const& value)
-{
-    j = std::string{to_cstring(value)};
-}
-
-//---------------------------------------------------------------------------//
-}  // namespace celeritas
 
 namespace celeritas
 {
