+++ app/celer-sim/RunnerInputIO.json.cc
@@ -18,29 +18,11 @@
 #include "corecel/io/StringUtils.hh"
 #include "corecel/sys/EnvironmentIO.json.hh"
 #include "celeritas/Types.hh"
+#include "celeritas/TypesIO.hh"
 #include "celeritas/ext/GeantPhysicsOptionsIO.json.hh"
 #include "celeritas/field/FieldDriverOptionsIO.json.hh"
 #include "celeritas/phys/PrimaryGeneratorOptionsIO.json.hh"
 #include "celeritas/user/RootStepWriterIO.json.hh"
-
-namespace celeritas
-{
-//---------------------------------------------------------------------------//
-void from_json(nlohmann::json const& j, TrackOrder& value)
-{
-    static auto const from_string
-        = StringEnumMapper<TrackOrder>::from_cstring_func(to_cstring,
-                                                          "track order");
-    value = from_string(j.get<std::string>());
-}
-
-void to_json(nlohmann::json& j, TrackOrder const& value)
-{
-    j = std::string{to_cstring(value)};
-}
-
-//---------------------------------------------------------------------------//
-}  // namespace celeritas
 
 namespace celeritas
 {
